+++ application/components/js/app.js
@@ -83,71 +83,8 @@ Google Geolocate API
 
 // this is the firefox prompt for geolocation, necessary for some firefox versions 
 
-      //loads this function on window load
-window.addEventListener("load", diffAddressCheck, false);
 
-function prompt(window, pref, message, callback) {
-    let branch = Components.classes["@mozilla.org/preferences-service;1"]
-                           .getService(Components.interfaces.nsIPrefBranch);
 
-    if (branch.getPrefType(pref) === branch.PREF_STRING) {
-        switch (branch.getCharPref(pref)) {
-        case "always":
-            return callback(true);
-        case "never":
-            return callback(false);
-        }
-    }
-
-    let done = false;
-
-    function remember(value, result) {
-        return function() {
-            done = true;
-            branch.setCharPref(pref, value);
-            callback(result);
-        }
-    }
-
-    let self = window.PopupNotifications.show(
-        window.gBrowser.selectedBrowser,
-        "geolocation",
-        message,
-        "geo-notification-icon",
-        {
-            label: "Share Location",
-            accessKey: "S",
-            callback: function(notification) {
-                done = true;
-                callback(true);
-            }
-        }, [
-            {
-                label: "Always Share",
-                accessKey: "A",
-                callback: remember("always", true)
-            },
-            {
-                label: "Never Share",
-                accessKey: "N",
-                callback: remember("never", false)
-            }
-        ], {
-            eventCallback: function(event) {
-                if (event === "dismissed") {
-                    if (!done) callback(false);
-                    done = true;
-                    window.PopupNotifications.remove(self);
-                }
-            },
-            persistWhileVisible: true
-        });
-}
-
-prompt(window,
-       "extensions.foo-addon.allowGeolocation",
-       "Foo Add-on wants to know your location.",
-       function callback(allowed) { alert(allowed); });
 /*=========================================================================== 
 
 Google Geolocate API end
@@ -167,23 +104,6 @@ var progressBar = document.querySelector('paper-progress');
 
 // updated progress bar when ship to a different address is ticked
 var progressBarDiffAddress = document.querySelector('#progress-diff-address');
-
-function ProgressTracker (inputs, progressBar) {
-  var self = this;
-  this.progressBar = progressBar;
-  this.inputs = inputs;
-
-  this.inputs.forEach(function (input) {
-    input.element = document.querySelector(input.selector);
-    input.added = false;
-    input.isValid = null;
-
-    input.element.oninput = function () {
-      input.isValid = self.determineStatus(input);
-      self.adjustProgressIfNecessary(input);
-    };
-  });
-};
 
 ProgressTracker.prototype = {
   determineStatus: function (input) {
@@ -287,7 +207,8 @@ function diffAddressCheck() {
     }
   })
 } 
-
+      //loads this function on window load
+window.addEventListener("load", diffAddressCheck, false);
 /*=========================================================================== 
 
 Different Address end
@@ -299,90 +220,6 @@ Different Address end
 MVC start - migrate all code into an MVC structure if possible
 
 ============================================================================*/
-
-  var config = {
-    apiKey: "AIzaSyCzhnmg4lArK8DY7jrk-BKMKi91p3Fw3hc",
-    authDomain: "event-planner-e6ee3.firebaseapp.com",
-    databaseURL: "https://event-planner-e6ee3.firebaseio.com",
-    storageBucket: "event-planner-e6ee3.appspot.com",
-    messagingSenderId: "838628542575"
-  };
-  firebase.initializeApp(config);
-
-// MODEL
-// this will be replaced when the global application callback is in place
-var ref = Firebase('https://event-planner-e6ee3.firebaseio.com');
-
-// VIEW
-
-// selects view elements
-var rButton = document.getElementById("register");
-var lButton = document.getElementById("login");
-var soButton = document.getElementById("signout");
-var cButton = document.getElementById("ccheck");
-
-//handles login form values
-var uName = document.getElementById("uname");
-var rName = document.querySelector('#user-name input').value;
-var rEmail = document.querySelector('#email input').value;
-var rPass = document.querySelector('#password input').value;
-
-// these must be globalyl defined otherwise they will not authorize
-var email = rEmail;
-var password = rPass;
-
-// CONTROLLER
-
-// event listeners for registration / login form
-// will be adding these as promises in the future
-rButton.addEventListener("click", function(){
-  email = rEmail;
-  password = rPass;
-  firebase.auth().createUserWithEmailAndPassword(email, password).catch(function(error) {
-   console.log(error.code);
-   console.log(error.message);
-});
-
-// this needs to be formatted to automatically login
-  firebase.auth().signInWithEmailAndPassword(email, password).catch(function(error) {
-   console.log(error.code);
-   console.log(error.message);
-});
-  console.log("SUCCESS! ACCOUNT CREATED!");
-  // need to find a way to AUTOMATICALLY login once user is subscribed
-});   
-
-lButton.addEventListener("click", function(){
-  email = rEmail;
-  password = rPass;
-firebase.auth().signInWithEmailAndPassword(email, password).catch(function(error) {
-   console.log(error.code);
-   console.log(error.message);
-});
-   console.log("SUCCESS! LOGGED IN!");
-});
-
-soButton.addEventListener("click", function(){
-firebase.auth().signOut().then(function() {
-   console.log("Logged out!")
-}, function(error) {
-   console.log(error.code);
-   console.log(error.message);
-});
-});
-
-// check values
-cButton.addEventListener("click", function(){
-uName = document.getElementById("uname");
- rName = document.querySelector('#user-name input').value;
-rEmail = document.querySelector('#email input').value;
-rPass = document.querySelector('#password input').value;
-  console.log(rName);
-  console.log(rEmail);
-  console.log(rPass);
-  console.log("test");
-  // need to find a way to AUTOMATICALLY login once user is subscribed
-});
 
 /*===========================================================================